test/Makefile.am: Fix build with builddir neq srcdir
Header files from src/ are not copied to builddir, and we need to
explicitly state to use the srcdir as prefix, otherwise builddir is used
as CWD and the headers are not found.

For LDADD, we want the build-time gnerated object files, so we want to
take them from builddir.